fix(#7410): generate decap-cms-core's typings from source
decap-cms-core shipped a 661-line hand-written index.d.ts that lived outside
`src/`, so the root tsconfig (`include: ["**/src/**/*"]`) never compiled it and
nothing checked it against the implementation. It had drifted: the registry
exposes 27 members, the declared `CMS` interface described 20. A consumer
written against the published typings gets errors on `getWidgets`,
`getEventListeners`, `removeEventListener`, `invokeEvent`, `getCustomFormats`,
`getCustomFormatsExtensions` and `getCustomFormatsFormatters` — all of which
work at runtime — plus `registerWidget`'s array form and its 4th `schema`
argument, which is what #7410 reported.

The fix is to stop hand-maintaining the contract:

- index.d.ts becomes src/types/index.ts, a real module rather than an ambient
  `declare module`, so it is compiled with everything else.
- registry, EditorComponent and the two index files become TypeScript.
- `DecapCmsCore` uses `satisfies CMS` instead of a `: CMS` annotation. An
  annotation only proves the implementation satisfies the interface; it cannot
  catch an interface that omits an API. With `satisfies`, the emitted type is
  derived from the implementation and `CMS` becomes a compatibility check.
- decap-cms-core publishes `dist/esm/index.d.ts`, emitted by
  `pnpm run build:types`.

Declaration emit needs project references, but only along the chain core
actually imports (lib-util, lib-widgets, ui-default), so this adds four
tsconfig.build.json files rather than one per package. The root tsconfig's
`declaration: true` was a no-op next to `noEmit: true` and made TS check
declaration portability for packages that publish none; turning it off there
removes that noise without losing anything.

decap-cms-app keeps its small ambient index.d.ts — its 30 imports come from a
side-effect `import './extensions.js'`, none of which reach its public type
surface — but it now re-exports decap-cms-core so config types resolve through
it.

Corrections that fell out, each previously invisible:

- GetAssetFunction declared `(asset: string) => {url, path, field?, fileObj}`;
  boundGetAsset is `(path, field?) => AssetProxy`. The one component using it
  had annotated its parameters `any`, which hid both the arity and the return
  type.
- CmsEventListener.handler is called with a second `options` argument.
- registerPreviewStyle's options are optional, so `raw` is too.
- registerMediaLibrary stores `options` on the registered library.
- createEditorComponent converts `fields` to Immutable and may leave
  `toPreview` as `false`; named that shape RegisteredEditorComponent instead of
  claiming the input type back.
- lib-auth's shim declared NetlifyAuthenticator.refresh with one parameter; the
  implementation takes (options, cb).

scripts/types-fixture type-checks the built declarations the way an external
consumer would, and runs in CI after the demo build. Without it nothing stops
the next drift: it fails if any of the above regresses.
