+++ package.json
@@ -2,13 +2,14 @@
   "scripts": {
     "start": "pnpm run develop",
     "develop": "nx run-many -t develop --parallel=999 --output-style=stream --exclude=decap-server",
-    "build": "pnpm run build:esm && nx run-many -t build",
+    "build": "pnpm run build:esm && pnpm run build:types && nx run-many -t build",
     "build:esm": "nx run-many -t build:esm --output-style=stream",
+    "build:types": "tsc --build packages/decap-cms-core/tsconfig.build.json",
     "build:demo": "pnpm run build && ncp packages/decap-cms/dist dev-test/dist/",
     "build-preview": "pnpm run build && nx run decap-cms:build-preview --output-style=stream",
     "type-check": "tsc --noEmit",
     "type-check:watch": "pnpm run type-check -- --watch",
-    "clean": "rimraf \"packages/*/dist\" dev-test/dist \"packages/*/node_modules\" \".nx/cache\"",
+    "clean": "rimraf \"packages/*/dist\" dev-test/dist \"packages/*/node_modules\" \".nx/cache\" \"packages/*/tsconfig.build.tsbuildinfo\"",
     "reset": "pnpm run clean",
     "test": "pnpm run lint && pnpm run type-check && pnpm run test:cypress-runner && pnpm run test:unit",
     "test:all": "pnpm run test && pnpm run test:e2e",
@@ -36,6 +37,7 @@
     "format:prettier": "prettier \"{{packages,scripts}/**/,}*.{js,jsx,ts,tsx,css}\"",
     "prepare": "husky install",
     "publish:packages": "pnpm publish -r --no-git-checks --report-summary",
+    "test:types": "pnpm run build:types && pnpm --filter decap-cms-types-fixture run test:types",
     "test:package-integrity": "node scripts/test-package-integrity.mjs"
   },
   "browserslist": [
