+++ packages/decap-cms-app/src/index.ts
@@ -1,6 +1,15 @@
 import { DecapCmsCore as CMS } from 'decap-cms-core';
 import './extensions.js';
 
+import type { CMS as CmsInterface } from 'decap-cms-core';
+
+export * from 'decap-cms-core';
+
+declare global {
+  // This is not really a global but a literal injected by babel
+  const DECAP_CMS_APP_VERSION: string;
+}
+
 // Log version
 if (typeof window !== 'undefined') {
   if (typeof DECAP_CMS_APP_VERSION === 'string') {
@@ -8,7 +17,7 @@ if (typeof window !== 'undefined') {
   }
 }
 
-export const DecapCmsApp = {
+export const DecapCmsApp: CmsInterface = {
   ...CMS,
 };
 export default CMS;
