+++ packages/decap-cms-backend-github/src/implementation.tsx
@@ -142,7 +142,7 @@ export default class GitHub implements Implementation {
   }
 
   async status() {
-    const api = await fetch(GITHUB_STATUS_ENDPOINT)
+    const api = await fetch(GITHUB_STATUS_ENDPOINT, undefined)
       .then(res => res.json())
       .then(res => {
         return res['components']
