+++ packages/decap-cms-backend-test/src/implementation.ts
@@ -26,7 +26,7 @@ import type {
   DataFile,
 } from 'decap-cms-lib-util';
 
-type RepoFile = { path: string; content: string | AssetProxy };
+type RepoFile = { path: string; content?: string | AssetProxy };
 type RepoTree = { [key: string]: RepoFile | RepoTree };
 
 type Diff = {
@@ -56,7 +56,7 @@ declare global {
 }
 
 window.repoFiles = window.repoFiles || {};
-window.repoFilesUnpublished = window.repoFilesUnpublished || [];
+window.repoFilesUnpublished = window.repoFilesUnpublished || {};
 window.repoNotes = window.repoNotes || {};
 
 function getFile(path: string, tree: RepoTree) {
@@ -97,7 +97,7 @@ function moveFile(path: string, newPath: string, tree: RepoTree, hasSubfolders: 
   );
   files.forEach(file => {
     deleteFile(file.path, tree);
-    writeFile(file.path.replace(sourceDir, destDir), file.content, tree);
+    writeFile(file.path.replace(sourceDir, destDir), file.content!, tree);
   });
 }
