--- packages/decap-cms-core/src/index.js
@@ -1,8 +1,0 @@
-import bootstrap from './bootstrap';
-import Registry from './lib/registry';
-
-export const DecapCmsCore = {
-  ...Registry,
-  init: bootstrap,
-};
-export default DecapCmsCore;
