+++ packages/decap-cms-core/src/lib/stega.ts
@@ -3,7 +3,7 @@ import { vercelStegaEncode } from '@vercel/stega';
 import { isImmutableMap, isImmutableList } from '../types/immutable';
 
 import type { Map as ImmutableMap, List } from 'immutable';
-import type { CmsField } from 'decap-cms-core';
+import type { CmsField } from '../types';
 
 /**
  * Context passed to encode functions, containing the current state of the encoding process
