+++ packages/decap-cms-core/src/reducers/entries.ts
@@ -66,6 +66,7 @@ import type {
   EntriesGroupRequestPayload,
   EntriesGroupFailurePayload,
   GroupOfEntries,
+  EntryGroup,
 } from '../types/redux';
 import type { StaticallyTypedRecord } from '../types/immutable';
 
@@ -463,7 +464,7 @@ export function selectEntries(state: Entries, collection: Collection) {
   return entries;
 }
 
-function getGroup(entry: EntryMap, selectedGroup: GroupMap) {
+function getGroup(entry: EntryMap, selectedGroup: GroupMap): EntryGroup {
   const label = selectedGroup.get('label');
   const field = selectedGroup.get('field');
 
@@ -472,7 +473,7 @@ function getGroup(entry: EntryMap, selectedGroup: GroupMap) {
     return {
       id: 'missing_value',
       label,
-      value: fieldData,
+      value: undefined,
     };
   }
