+++ packages/decap-cms-core/src/valueObjects/EditorComponent.ts
@@ -1,13 +1,17 @@
 import { fromJS } from 'immutable';
 import isFunction from 'lodash/isFunction';
 
+import type { EditorComponentOptions, RegisteredEditorComponent } from '../types';
+
 const catchesNothing = /.^/;
 
-function bind(fn) {
+function bind(fn: unknown) {
   return isFunction(fn) && fn.bind(null);
 }
 
-export default function createEditorComponent(config) {
+export default function createEditorComponent(
+  config: EditorComponentOptions,
+): RegisteredEditorComponent {
   const {
     id = null,
     label = 'unnamed component',
@@ -32,7 +36,7 @@ export default function createEditorComponent(config) {
     fromBlock: bind(fromBlock) || (() => ({})),
     toBlock: bind(toBlock) || (() => 'Plugin'),
     toPreview: bind(toPreview) || (!widget && (bind(toBlock) || (() => 'Plugin'))),
-    fields: fromJS(fields),
+    fields: fromJS(fields) as RegisteredEditorComponent['fields'],
     ...remainingConfig,
   };
 }
