+++ packages/decap-cms-editor-component-image/package.json
@@ -16,7 +16,7 @@
   "scripts": {
     "develop": "pnpm run build:esm --watch",
     "build": "cross-env NODE_ENV=production webpack",
-    "build:esm": "cross-env NODE_ENV=esm babel src --out-dir dist/esm --ignore \"**/__tests__\" --root-mode upward"
+    "build:esm": "cross-env NODE_ENV=esm babel src --out-dir dist/esm --ignore \"**/__tests__\" --root-mode upward --extensions \".js,.jsx,.ts,.tsx\""
   },
   "devDependencies": {
     "cross-env": "catalog:"
