+++ packages/decap-cms-lib-auth/index.d.ts
@@ -2,10 +2,11 @@ declare module 'decap-cms-lib-auth' {
   class NetlifyAuthenticator {
     constructor(config = {});
 
-    refresh: (args: {
-      provider: string;
-      refresh_token: string;
-    }) => Promise<{ token: string; refresh_token: string }>;
+    // src/netlify-auth.js: refresh(options, cb)
+    refresh: (
+      args: { provider: string; refresh_token: string },
+      cb?: (error: Error | null, data?: unknown) => void,
+    ) => Promise<{ token: string; refresh_token: string }>;
   }
 
   class PkceAuthenticator {
