+++ pnpm-workspace.yaml
@@ -1,5 +1,6 @@
 packages:
   - packages/*
+  - scripts/types-fixture
 
 catalog:
   '@babel/cli': ^7.11.6
@@ -251,7 +252,7 @@ catalog:
   ts-loader: ^9.4.2
   ts-node: ^10.9.1
   tsconfig-paths-webpack-plugin: ^3.2.0
-  typescript: ^4.9.4
+  typescript: ^5.3.0
   unified: ^9.2.0
   unist-builder: ^1.0.3
   unist-util-visit: ^1.4.0
