+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "declaration": true,
+    "declaration": false,
     "jsx": "react-jsx",
     "jsxImportSource": "@emotion/react",
     "target": "esnext",
@@ -17,7 +17,7 @@
     "useUnknownInCatchVariables": false,
     "baseUrl": "./",
     "paths": {
-      "decap-cms-core": ["packages/decap-cms-core"],
+      "decap-cms-core": ["packages/decap-cms-core/src"],
       "decap-cms-backend-bitbucket": ["packages/decap-cms-backend-bitbucket/src"],
       "decap-cms-backend-github": ["packages/decap-cms-backend-github/src"],
       "decap-cms-backend-gitlab": ["packages/decap-cms-backend-gitlab/src"],
